Added a health system
Added a bar in the top left corner of the game to showcase the player's current health.
Added a stationary spike enemy in the test environment so you can test the damage system.
The health bar decreases, when the player touches an enemy.
Each character has a HurtBox and a HitBox. The HitBox needs to touch the other character's HurtBox to inflict damage.