SCRUM-46 | Added a turn based combat system
Added a system, where a player can choose to deal damage to the enemy or heal himself through clicking on the "cards". Right now its simple buttons, but in the future need to change it to cards
Need to finish implementing what happens when the player's health reaches 0.
Also need to implement a victory screen or something that pops up when you kill an enemy
The player needs to click "End turn" button to end his turn
The enemy can choose from two options when it is his turn. He can either attack for 1 damage or heal himself for one point. Each option has a 50/50 chance. The enemy performs two actions for his turn
Added code to other scripts for better readability
Fixed the transition screen, to clear himself after the animation, since the transition screen blocked mouse inputs.